Convert some parens initializers to curly brace initializers

+++ src/server/shell/mouse_keys_transformer.h
@@ -67,11 +67,11 @@ struct MouseKeysTransformer: public mir::input::InputEventTransformer::Transform
     {
         // Quadratic, linear, and constant factors repsectively
         // ax^2 + bx + c
-        float const a, b, c;
+        double const a, b, c;
 
         AccelerationCurve(std::shared_ptr<mir::options::Option> const& options);
 
-        float evaluate(float t) const;
+        double evaluate(double t) const;
     };
 
     std::shared_ptr<mir::MainLoop> const main_loop;
